+++ core/server/api_container/server/startosis_engine/startosis_interpreter.go
@@ -16,7 +16,6 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_errors"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_modules"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_modules/proto_compiler"
-	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	starlarkproto "go.starlark.net/lib/proto"
 	"go.starlark.net/lib/time"
@@ -133,14 +132,14 @@ func (interpreter *StartosisInterpreter) addInputArgsToPredeclared(moduleId stri
 
 	// Get descriptor for type "ModuleInput" in the module types.proto file
 	protoTypesFile := strings.Join([]string{moduleId, TypesFileName}, string(filepath.Separator))
-	fileStore, err := interpreter.protoFileStore.LoadProtoFile(protoTypesFile)
-	if err != nil && serializedJsonParams == EmptyInputArgs {
+	fileStore, interpretationError := interpreter.protoFileStore.LoadProtoFile(protoTypesFile)
+	if interpretationError != nil && serializedJsonParams == EmptyInputArgs {
 		// If am empty param was passed to the script, then it's valid to not have a types.proto inside the module
 		(*predeclared)[MainInputArgName] = starlark.None
 		return nil
 	}
-	if err != nil {
-		return startosis_errors.NewInterpretationError("File '%s' not found at the root of module '%s' but a non empty parameter was passed. This is allowed to define a module with no '%s', but it should be always be called with an empty parameter", TypesFileName, moduleId, TypesFileName)
+	if interpretationError != nil {
+		return startosis_errors.WrapError(interpretationError, "File '%s' either absent of invalid at the root of module '%s' but a non empty parameter was passed. This is allowed to define a module with no '%s', but it should be always be called with an empty parameter", TypesFileName, moduleId, TypesFileName)
 	}
 	reflectMessageDescriptor, err := fileStore.FindDescriptorByName(ModuleInputTypeName)
 	if err != nil && serializedJsonParams == EmptyInputArgs {
@@ -149,7 +148,7 @@ func (interpreter *StartosisInterpreter) addInputArgsToPredeclared(moduleId stri
 		return nil
 	}
 	if err != nil {
-		return startosis_errors.NewInterpretationError("Type '%s' cannot be found in type file '%s' for module '%s' but a non empty parameter was passed. When some parameters are passed to a module, there must be a `%s` type defined in the module's '%s' file", ModuleInputTypeName, TypesFileName, moduleId, MainInputArgName, ModuleInputTypeName)
+		return startosis_errors.WrapError(err, "Type '%s' cannot be found in type file '%s' for module '%s' but a non empty parameter was passed. When some parameters are passed to a module, there must be a `%s` type defined in the module's '%s' file", ModuleInputTypeName, TypesFileName, moduleId, MainInputArgName, ModuleInputTypeName)
 	}
 	messageDescriptor, ok := reflectMessageDescriptor.(protoreflect.MessageDescriptor)
 	if !ok {
@@ -167,13 +166,11 @@ func (interpreter *StartosisInterpreter) addInputArgsToPredeclared(moduleId stri
 	// Convert the proto.Message into a starlarkproto.Message
 	protobufMarshalledMessage, err := proto.Marshal(message)
 	if err != nil {
-		logrus.Error(stacktrace.Propagate(err, "Unable to marshal proto message '%s' from module ID '%s'", ModuleInputTypeName, moduleId).Error())
-		return startosis_errors.NewInterpretationError("Unable to serialize the '%s' type of module '%s'. This is unexpected. More info will be logged to Kurtosis core", ModuleInputTypeName, moduleId)
+		return startosis_errors.WrapError(err, "Unable to serialize the '%s' type of module '%s'. This is unexpected. More info will be logged to Kurtosis core", ModuleInputTypeName, moduleId)
 	}
 	starlarkMessage, err := starlarkproto.Unmarshal(message.Descriptor(), protobufMarshalledMessage)
 	if err != nil {
-		logrus.Error(stacktrace.Propagate(err, "Unable to convert proto message '%s' to a starlark proto message from module ID '%s'", ModuleInputTypeName, moduleId).Error())
-		return startosis_errors.NewInterpretationError("Unable to serialize the '%s' type of module '%s'. This is unexpected. More info will be logged to Kurtosis core", moduleId, serializedJsonParams)
+		return startosis_errors.WrapError(err, "Unable to serialize the '%s' type of module '%s'. This is unexpected. More info will be logged to Kurtosis core", moduleId, serializedJsonParams)
 	}
 	(*predeclared)[MainInputArgName] = starlarkMessage
 	return nil
@@ -219,8 +216,8 @@ func (interpreter *StartosisInterpreter) makeLoadFunction(instructionsQueue *[]k
 		}()
 
 		// Load it.
-		contents, err := interpreter.moduleContentProvider.GetModuleContents(moduleID)
-		if err != nil {
+		contents, interpretationError := interpreter.moduleContentProvider.GetModuleContents(moduleID)
+		if interpretationError != nil {
 			return nil, startosis_errors.NewInterpretationError("An error occurred while loading the module '%v'", moduleID)
 		}