+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - Prints out the instruction line, col & filename in the execution error
 - Prints out the instruction line, col & filename in the validation error
 - Added `remove_service` to Startosis
+- InterpretationError is now able to store a `cause`. It simplifies being more explicit on want the root issue was
 
 # 0.51.7